Adjusted DynSearchService.java to use the json confic as default
